refactor: https://github.com/next-step/java-bowling/pull/262#discussion_r485694574 수정 반영
