+++ src/main/java/bowling/Main.java
@@ -1,20 +1,18 @@
 package bowling;
 
-import bowling.domain.FrameSet;
+import bowling.domain.Bowler;
 import bowling.view.InputScanner;
 import bowling.view.Printer;
 
 public class Main {
     public static void main(String[] args) {
-        String bowler = InputScanner.getBowler("플레이어 이름은(3 english letters)?: ");
+        Bowler bowler = InputScanner.getBowler("플레이어 이름은(3 english letters)?: ");
 
-        FrameSet frameSet = FrameSet.from(bowler);
+        while (bowler.isPlaying()) {
+            int hitCount = InputScanner.getHitCount(String.format("%s프레임 투구 : ", bowler.getPlayStage()));
+            bowler.record(hitCount);
 
-        while (frameSet.isPlaying()) {
-            int hitCount = InputScanner.getHitCount(String.format("%s프레임 투구 : ", frameSet.getPlayStage()));
-            frameSet.record(hitCount);
-
-            Printer.showResult(frameSet);
+            Printer.showResult(bowler);
         }
     }
 }
