+++ src/main/java/bowling/domain/Bowler.java
@@ -1,17 +1,36 @@
 package bowling.domain;
 
+import java.util.List;
 import java.util.Objects;
 
 public class Bowler {
 
     private String name;
+    private Frames frames;
 
     private Bowler(String name) {
         this.name = name;
+        frames = Frames.create();
     }
 
     public static Bowler from(String name) {
         return new Bowler(name);
+    }
+
+    public boolean isPlaying() {
+        return frames.isPlaying();
+    }
+
+    public void record(int hitCount) {
+        frames.record(hitCount);
+    }
+
+    public int getPlayStage() {
+        return frames.getPlayStage();
+    }
+
+    public List<Frame> getFrames() {
+        return frames.getFrames();
     }
 
     @Override
