+++ src/main/java/bowling/view/InputScanner.java
@@ -1,15 +1,17 @@
 package bowling.view;
 
+import bowling.domain.Bowler;
+
 import java.util.Scanner;
 
 public class InputScanner {
     private static final Scanner scanner = new Scanner(System.in);
 
-    public static String getBowler(String message) {
+    public static Bowler getBowler(String message) {
         System.out.print(message);
         String result = scanner.next();
         System.out.println();
-        return result;
+        return Bowler.from(result);
     }
 
     public static int getHitCount(String message) {
