+++ src/main/java/bowling/view/Printer.java
@@ -1,30 +1,30 @@
 package bowling.view;
 
+import bowling.domain.Bowler;
 import bowling.domain.Frame;
-import bowling.domain.FrameSet;
 import bowling.domain.GameResults;
 
 import java.util.stream.Collectors;
 
 public class Printer {
-    public static void showResult(FrameSet frameSet) {
-        showHead(frameSet);
-        showBody(frameSet);
+    public static void showResult(Bowler bowler) {
+        showHead(bowler);
+        showBody(bowler);
         System.out.println();
     }
 
-    private static void showHead(FrameSet frameSet) {
+    private static void showHead(Bowler bowler) {
         System.out.print("| NAME |");
-        for (Frame frame : frameSet.getFrames()) {
+        for (Frame frame : bowler.getFrames()) {
             System.out.print(String.format("  %02d  |", frame.getStage()));
         }
         System.out.println();
     }
 
-    private static void showBody(FrameSet frameSet) {
-        System.out.print(String.format("| %4s |", frameSet.getBowler()));
+    private static void showBody(Bowler bowler) {
+        System.out.print(String.format("| %4s |", bowler.toString()));
 
-        for (Frame frame : frameSet.getFrames()) {
+        for (Frame frame : bowler.getFrames()) {
             GameResults results = frame.getResults();
             System.out.printf(
                     String.format("  %-3s |",
